Move onboarding background to desktop shell
