+++ lib/src/features/onboarding/create/onboarding_split_view.dart
@@ -134,6 +134,7 @@ class OnboardingSplitViewShell extends ConsumerWidget {
 
     return AppDesktopShell(
       sidebarWidth: 256,
+      background: _OnboardingWindowBackground(activeStep: activeStep),
       sidebar: SlideTransition(
         position: sidebarOffset,
         child: _Sidebar(
@@ -143,6 +144,37 @@ class OnboardingSplitViewShell extends ConsumerWidget {
         ),
       ),
       pane: FadeTransition(opacity: entrance, child: child),
+    );
+  }
+}
+
+class _OnboardingWindowBackground extends StatelessWidget {
+  const _OnboardingWindowBackground({required this.activeStep});
+
+  final OnboardingStep activeStep;
+
+  @override
+  Widget build(BuildContext context) {
+    final isDark = AppTheme.of(context) == AppThemeData.dark;
+    final asset = switch (activeStep) {
+      OnboardingStep.secretPassphrase =>
+        isDark
+            ? 'assets/illustrations/onboarding_secret_passphrase_background_dark.png'
+            : 'assets/illustrations/onboarding_secret_passphrase_background_light.png',
+      _ => null,
+    };
+
+    if (asset == null) {
+      return const SizedBox.shrink();
+    }
+
+    return DecoratedBox(
+      decoration: BoxDecoration(color: context.colors.background.window),
+      child: Image.asset(
+        asset,
+        fit: BoxFit.cover,
+        alignment: Alignment.topCenter,
+      ),
     );
   }
 }
